% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompaniesPoland.R
\docType{data}
\name{CompaniesPoland}
\alias{CompaniesPoland}
\title{Exemplary data for calculating an Index of Spatial agglomeration}
\format{CompaniesPoland is a data frame with 13 rows (companies) and 4 columns (company parameters) named lon and lat 
(geographic longitude and latitude of the company), employment and category.}
\usage{
CompaniesPoland
}
\description{
This data frame contains a random set of companies located across Poland. Every company has an artificial
category assigned, which defines the field of work the company focuses on and the number of employees working
there.
}
\keyword{datasets}

