% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPoland.R
\docType{data}
\name{MapPoland}
\alias{MapPoland}
\title{Exemplary map for calculating an Index of Spatial agglomeration}
\format{\code{MapPoland} is a \code{SpatialPolygonsDataFrame} object representing a map of Poland.}
\source{
\href{http://www.codgik.gov.pl/}{\code{Centralny Osrodek Dokumentacji Geodezyjnej i Kartograficznej}}
}
\usage{
MapPoland
}
\description{
This object is a \code{SpatialPolygonsDataFrame} obtained via using \code{readOGR} function from rgdal package on a shapefile.
The data frame represents a map of Poland divided into voivodeships.
}
\seealso{
\code{\link{SpatialPolygonsDataFrame}}
}
\keyword{datasets}
