% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapefilePoland.R
\docType{data}
\name{ShapefilePoland}
\alias{ShapefilePoland}
\title{Exemplary map for calculating an Index of Spatial agglomeration}
\format{ShapefilePoland is a \code{SpatialPolygonsDataFrame} object representing a map of Poland.}
\usage{
ShapefilePoland
}
\description{
This object is a \code{SpatialPolygonsDataFrame} obtained via using \code{readOGR} function from rgdal package on a shapefile.
The data frame represents a map of Poland divided into voivodeships.
}
\seealso{
\code{\link{SpatialPolygonsDataFrame}}
}
\keyword{datasets}

