% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAG.R
\name{SPAG}
\alias{SPAG}
\title{Function calculating the coverage, distance and overlap components of the SPAG Index.}
\usage{
SPAG(companiesDF, xInd = 1, yInd = 2, empInd = 3, categInd = 4, shp)
}
\arguments{
\item{companiesDF}{- data frame with information regarding the companies. At least four columns are required:
x and y coordinates of the company, the category of company and a numeric}

\item{xInd}{- number of the column in companiesDF with information regarding the latitude}

\item{yInd}{- number of the column in companiesDF with information regarding the longitude}

\item{empInd}{- number of the column in companiesDF with numeric data regarding the employment}

\item{categInd}{- number of the column in companiesDF information about the category of the company}

\item{shp}{- SpatialPolygonsDataFrame object obtained via loading a shapefile}
}
\description{
Function calculating the coverage, distance and overlap components of the SPAG Index.
}

