% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAG.R
\name{SPAG}
\alias{SPAG}
\title{Function calculating the coverage, distance and overlap components of the SPAG Index.}
\usage{
SPAG(companiesDF, shp, xInd = 1, yInd = 2, empInd = 3, categInd = 4)
}
\arguments{
\item{companiesDF}{- data frame with information regarding the companies. At least four columns are required:
x and y coordinates of the company, the category of company and a numeric}

\item{shp}{- SpatialPolygonsDataFrame object obtained via loading a shapefile}

\item{xInd}{- number of the column in companiesDF with information regarding the latitude}

\item{yInd}{- number of the column in companiesDF with information regarding the longitude}

\item{empInd}{- number of the column in companiesDF with numeric data regarding the employment}

\item{categInd}{- number of the column in companiesDF information about the category of the company}
}
\description{
Function calculating the coverage, distance and overlap components of the SPAG Index.
}
\examples{
data(CompaniesPoland)
data(ShapefilePoland)
spagIndex <- SPAG(CompaniesPoland,shp = ShapefilePoland)
print(spagIndex)
plot(spagIndex)
plot(spagIndex, addCompanies = FALSE)
plot(spagIndex, category = "A")
plot(spagIndex, category = "A") + coord_map("orthographic")
plot(spagIndex, category = "B")
plot(spagIndex, category = "C")

}
